fix!: rewrite the `addTimeoutToPromise` helper (#268)
During testing it turns out that in some places the previous implementation
did not work correctly, probably due to some bugs in ALS. To get around those,
we now use the same way as before with a regular `setTimeout/clearTimeout` for
the timeout handler. We no longer have `cancelTimeout` abort controller instance,
as we clear the timeout instead of using abort signal.