feat: allow custom proxies to be https (#392)
Ondra needs to have custom proxy urls to allow https since he want's to
test some provider proxies and they do not have http. He says it's
working in crawlee, but our input schema does not allow it. I checked
and it looks like this regex is only used for that one check.

---------

Co-authored-by: Jaroslav Hejlek <jarda@hejlek.cz>