chore: update eslint to v9 (#508)
Disabled a few of the rules that would be too time-consuming to fix
(with dozens of errors for each), otherwise the changes are mostly from
the fixer, plus a few small final touches.

Also updated TS, which required one manual adjustment (and will unlock
the lockfile update PR from renovate).

Closes #489