feat: add general access enum to consts (#507)
We plan to allow users to configure through API & clients how their
storages and runs can be accessed — whether just the ID / name is
sufficient to access the storage. Full context in
https://github.com/apify/apify-core/issues/19012.

This PR adds appropriate enums that will be used both externally (in
clients & API), and internally (where it will replace the
[existing](https://github.com/apify/apify-core/blob/ba052eedd7490952fa1b8c1447c261a9cfd3acb3/src/packages/consts/src/iam.ts#L101)
enum).

Note that there are separate enums for runs and storages as only
storages support access by name. So far we have been using a shared enum
for both, but it's cleaner to have them separated — in run related logic
& documentation we won't have to account (much) for the unsupported
value.

---------

Co-authored-by: Martin Adámek <banan23@gmail.com>