feat: create injectable regex utility function (#536)
This PR introduces a new utility function, `createInjectableRegExp()`,
that converts any regex into an "injectable" version by removing its
leading `^` and trailing `$` anchors.

This is useful when you need to embed an existing pattern, which was
originally designed to match a whole string, inside another regex to
match it as a substring.

Thanks to this, we can replace
[APIFY_PROXY_VALUE_REGEX](https://github.com/apify/apify-shared-js/blob/master/packages/consts/src/regexs.ts#L51)
with
[REGEX_STR_USERNAME_VALUE](https://github.com/apify/apify-proxy/blob/develop/src/proxy_utils.ts#L20)
in Apify proxy and solve this issue
https://github.com/apify/apify-proxy/issues/1079.

The function was created in Apify shared repo, so it can be used in
other places.