ci: ensure JSON schemas are in sync (#539)
follow-up to https://github.com/apify/apify-shared-js/pull/538

We need to ensure that for schemas that have `.ts` file used to generate
`.json` file:
- users modify the `.ts` file not the `.json` file.
- users run `npm run build` to generate the `.json` file and commits
this generated file.
- that means -> `.ts` and `.json` files are always in sync

This adds CI workflow that runs whenever this `.ts` or `.json` file
changed, runs the build and check if `.json` file changes or not.