feat: add fileupload editor for string arrays (#545)
This PR allows the use of the `fileupload` input for an array of
strings.

Instead of re-using the `arrayItemsStringlist` reference, I've added
`arrayItemsFileupload` with the same shape.

It will work the same way as it currently works for a single string.

Related PR in console: https://github.com/apify/apify-core/pull/23289

The order of PRs does not matter: The UI will fallback to code input,
and the input just won't do anything until it encounters the editor in
schema.