feat: parse boolean utility function (#577)
This PR introduces simple new `parseBooleanOrNull` utility function to
parse boolean from string.
If value is stringbool, then the function return `true`/`false`.
if the value is invalid, then the function return `null`, which is I
think the best as we can distinguish between valid and invalid input,
but also null is falsy value, so we can use it as false in condition.

In many repos, we do not use Zod validator which can handle this using
coerce api, and it would be a very heavy dependency for just this
usecase.